Restricted Page: only host can update/delete room, redirect to home page if already login

+++ base/views.py
@@ -1,6 +1,7 @@
-from multiprocessing import context
 from django.shortcuts import render, redirect
+from django.http.response import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -18,6 +19,11 @@ from .forms import RoomForm
 
 
 def loginPage(request):
+
+    # if user already login, will redirect to home page when trying to enter login page
+    if request.user.is_authenticated:
+        return redirect("home")
+
     if request.method == "POST":
         # get value from frontend
         username = request.POST.get("username")
@@ -65,6 +71,7 @@ def room(request, pk):
     return render(request, "base/room.html", context)
 
 
+@login_required(login_url="login")
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -77,9 +84,14 @@ def createRoom(request):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def updateRoom(request, pk):  # pk: primiary key
     room = Room.objects.get(id=pk)  # get Room
     form = RoomForm(instance=room)  # add room into room form
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!")
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -90,8 +102,13 @@ def updateRoom(request, pk):  # pk: primiary key
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!")
+
     if request.method == "POST":
         room.delete()
         return redirect("home")
